requirements.txt
transformers
peft
datasets
diffusers
pillow
gradio
accelerate
bitsandbytes
torch
xformers